GUI portion
This is the GUI of the application mostly finished. It pretty much just needs the game logic now. The action listeners for the buttons are created in a for loop, and they should be accessible by the number on them. I still have to add the gif in the "0" square, but for now you can just use 0 as your reference for that.